Move `XOnlyPublicKey` to `key` module
