+++ src/lib.rs
@@ -156,6 +156,7 @@ pub use key::SecretKey;
 pub use key::PublicKey;
 pub use key::ONE_KEY;
 pub use key::KeyPair;
+pub use key::XOnlyPublicKey;
 pub use context::*;
 use core::marker::PhantomData;
 use core::{mem, fmt, ptr, str};
